move tests to appropriate place
