+++ src/helpers/network.rs
@@ -228,78 +228,6 @@ mod tests {
     }
 
     #[tokio::test]
-    async fn successfully_receives() {
-        let expected_message_chunks = (
-            ChannelId::new(Role::H2, Step::default().narrow("successfully-receives")),
-            vec![0u8; MESSAGE_PAYLOAD_SIZE_BYTES],
-        );
-        let expected_query_id = QueryId;
-
-        let transports = InMemoryNetwork::default();
-        let identities = transports.helper_identities();
-        let this_identity = identities[Role::H1];
-        let that_identity = identities[Role::H2];
-        let this_transport = transports.transport(this_identity).unwrap();
-        let this_network = Network::new(
-            this_transport.clone(),
-            expected_query_id,
-            RoleAssignment::new(identities),
-        );
-        let mut message_chunks_stream = this_network.recv_stream().await;
-
-        let that_transport = transports.transport(that_identity).unwrap();
-        let command = TransportCommand::StepData {
-            query_id: expected_query_id,
-            step: expected_message_chunks.0.step.clone(),
-            payload: expected_message_chunks.1.clone(),
-            offset: 0,
-        };
-        that_transport.send(this_identity, command).await.unwrap();
-
-        let message_chunks = message_chunks_stream.next().await;
-        assert_eq!(message_chunks, Some(expected_message_chunks));
-    }
-
-    #[tokio::test]
-    #[should_panic(expected = "NoSubscribers")]
-    async fn fails_if_not_subscribed() {
-        let expected_query_id = QueryId;
-        let message_chunks = (
-            ChannelId::new(Role::H2, Step::default().narrow("no-subscribe")),
-            vec![0u8; MESSAGE_PAYLOAD_SIZE_BYTES],
-        );
-
-        let transports = InMemoryNetwork::default();
-        let identities = transports.helper_identities();
-        let this_identity = identities[Role::H1];
-        let that_identity = identities[Role::H2];
-        let this_transport = transports.transport(this_identity).unwrap();
-        let this_network = Network::new(
-            this_transport.clone(),
-            expected_query_id,
-            RoleAssignment::new(identities),
-        );
-
-        let command = TransportCommand::StepData {
-            query_id: expected_query_id,
-            step: message_chunks.0.step.clone(),
-            payload: message_chunks.1.clone(),
-            offset: 0,
-        };
-
-        // missing:
-        // let mut message_chunks_stream = this_network.recv_stream().await;
-
-        let that_transport = transports.transport(that_identity).unwrap();
-        let _sent = that_transport.send(this_identity, command).await;
-
-        // TODO: there seems to be a race condition here?
-        // must await to drive the in-memory transport to the error?
-        let mut message_chunks_stream = this_network.recv_stream().await;
-        message_chunks_stream.next().await.unwrap();
-    }
-
-    #[tokio::test]
     #[should_panic(expected = "out-of-order delivery of data for")] // just the prefix
     async fn rejects_bad_offset() {
         let expected_query_id = QueryId;
