+++ src/test_fixture/transport/routing.rs
@@ -111,9 +111,7 @@ impl Switch {
                         }
                     }
                     Some((origin, command)) = peer_links.next() => {
-                        command.dispatch(CommandOrigin::Helper(origin), &routes)
-                        .await
-                        .unwrap_or_else(|err| panic!("Failed to dispatch a command: {err}"));
+                        command.dispatch(CommandOrigin::Helper(origin), &routes).await.expect("Failed to dispatch a command");
                     }
                     else => {
                         tracing::debug!("All channels are closed and switch is terminated");
